Remove support for syncing member custom fields and TJ tags
The existing tags being sync'ed (`rts` and `deceased`) should now
be set and searched for on Id only, so they don't need to be sync'ed
to/from TJ. Further, now that the active flag is being set on members
in Id via a lambda, the number of custom fields that aren't being
sync'd has increased 100%, causing days worth of extra processing. So
just remove this support, and build out anything further in the
tasks repo.

This does not remove support for `syncid` tags or pushing lists from
Id to a TJ tag.